refactored codes
janko in separate files, at gitignore
