--- utils_api.txt
@@ -1,13 +1,0 @@
-def reset_board(board):
-def min_coord(board):
-def is_deadend(board):
-def has_no_null(board):
-def is_legit(board, choices, diag):
-def cancel_all(board, constraint, choices, diag):
-def optimize(board, constraint, choices, coord, diag):
-def printBorder(maxChar, dim):
-def printOut(board):
-def mass_optimize(board, constraint, choices, diag):
-def init_board(constraint, choices, diag):
-def solve(shit_to_solve, constraint, choices, diag, solutions_list, counts):
-def solvable_without_trials(constraint, choices, diag):
